bencode: Correcting the data type for ByteStr length.

+++ bencode/bytestr.h
@@ -21,7 +21,7 @@ namespace BEncode {
 			ByteStr& operator=(const ByteStr&);
 
 			char* _buf;
-			uint64_t _len;
+			size_t _len;
 	};
 
 }
